Make collection field logic on BaseDocumentForm reusable
