+++ wagtail/wagtaildocs/forms.py
@@ -8,46 +8,13 @@ from django.template.loader import render_to_string
 from django.utils.translation import ugettext_lazy as _
 
 from wagtail.wagtailadmin import widgets
+from wagtail.wagtailadmin.forms import BaseCollectionMemberForm
 from wagtail.wagtailcore.models import Collection, GroupCollectionPermission
-from wagtail.wagtaildocs.permissions import permission_policy
+from wagtail.wagtaildocs.permissions import permission_policy as documents_permission_policy
 
 
-class BaseDocumentForm(forms.ModelForm):
-    def __init__(self, *args, **kwargs):
-        user = kwargs.pop('user', None)
-
-        super(BaseDocumentForm, self).__init__(*args, **kwargs)
-
-        if user is None:
-            self.collections = Collection.objects.all()
-        else:
-            self.collections = (
-                permission_policy.collections_user_has_permission_for(user, 'add')
-            )
-
-        if self.instance.pk:
-            # editing an existing document; ensure that the list of available collections
-            # includes its current collection
-            self.collections = (
-                self.collections | Collection.objects.filter(id=self.instance.collection_id)
-            )
-
-        if len(self.collections) == 0:
-            raise Exception(
-                "Cannot construct DocumentForm for a user with no document collection permissions"
-            )
-        elif len(self.collections) == 1:
-            # don't show collection field if only one collection is available
-            del self.fields['collection']
-        else:
-            self.fields['collection'].queryset = self.collections
-
-    def save(self, commit=True):
-        if len(self.collections) == 1:
-            # populate the instance's collection field with the one available collection
-            self.instance.collection = self.collections[0]
-
-        return super(BaseDocumentForm, self).save(commit=commit)
+class BaseDocumentForm(BaseCollectionMemberForm):
+    permission_policy = documents_permission_policy
 
 
 def get_document_form(model):
